requirements.txt
pyTelegramBotAPI
requests
python-dotenv
rapidfuzz
SpeechRecognition
pydub
librosa
soundfile
numpy